refactor: split git_overlay_health into sub-modules (#609 phase 2)
Phase 2 of the #609 god-module decomposition. Splits the second-largest
command file, git_overlay_health.rs (4013 LOC), by extracting its single
trailing `#[cfg(test)] mod tests` block (471 LOC) into a sibling tests.rs,
mirroring the proven Phase-1 command_catalog seam.

git_overlay_health.rs becomes a git_overlay_health/ module directory:
  - git_overlay_health/mod.rs    3544 LOC  production code (unchanged)
  - git_overlay_health/tests.rs   475 LOC  the mod tests body, moved verbatim

Pure behavior-preserving code movement:
- The test body moved unchanged (de-indented one level out of the
  `mod tests {}` wrapper). It referenced the parent via `use super::{...}`
  and `use crate::cli::commands::build_command_catalog`; both resolve
  identically from a sibling file (super -> the git_overlay_health module).
- The parent file's only delta is the inline `mod tests { ... }` block
  replaced by `#[cfg(test)] mod tests;`.
- No public item renamed, no signature changed, no logic touched.
- git detects an 88%-similar rename + the new tests.rs.

Refs #609. mount/core.rs and repository.rs remain for later phases.
